+++ CHANGELOG.md
@@ -3,10 +3,14 @@
 
 ## [0.7] - Unreleased
 
-## Added
+### Added
 
 - Added `hyphae`, a block syncing bridge between `myceli` and `kubo`.
 - Added `RequestVersion` API to `myceli`.
+
+### Changed
+
+- Added `mtu` config parameter to `myceli`.
 
 ## [0.6] - 2023-04-04